synchro avec les évènements phase dossier
